Refactoring - extract out logic that merges impacts within the same field
